+++ packages/api-reference/src/plugins/index.ts
@@ -1,7 +1,7 @@
 export {
   createPluginManager,
-  PLUGIN_MANAGER_SYMBOL,
   type ApiReferencePlugin,
   type PluginManager,
-  usePluginManager,
-} from './plugin-manager'
+} from '@/plugins/plugin-manager'
+
+export { PLUGIN_MANAGER_SYMBOL, usePluginManager } from '@/plugins/hooks/usePluginManager'
