+++ packages/api-reference/src/plugins/plugin-manager.ts
@@ -1,5 +1,4 @@
 import type { OpenApiExtension, ApiReferencePlugin as OriginalApiReferencePlugin } from '@scalar/types/api-reference'
-import { type InjectionKey, inject } from 'vue'
 
 export type ApiReferencePlugin = OriginalApiReferencePlugin
 
@@ -34,19 +33,3 @@ export const createPluginManager = ({ plugins = [] }: CreatePluginManagerParams)
 }
 
 export type PluginManager = ReturnType<typeof createPluginManager>
-export const PLUGIN_MANAGER_SYMBOL = Symbol() as InjectionKey<PluginManager>
-
-/**
- * Hook to access the plugin manager
- *
- * @throws Error if plugin manager is not provided
- */
-export const usePluginManager = (): PluginManager => {
-  const manager = inject(PLUGIN_MANAGER_SYMBOL)
-
-  if (!manager) {
-    throw new Error('Plugin manager not provided')
-  }
-
-  return manager
-}
